+++ examples/run_dien.py
@@ -22,21 +22,21 @@ def get_xy_fd(use_neg=False, hash_flag=False):
     uid = np.array([0, 1, 2])
     ugender = np.array([0, 1, 0])
     iid = np.array([1, 2, 3])  # 0 is mask value
-    igender = np.array([1, 2, 1])  # 0 is mask value
+    cate_id = np.array([1, 2, 2])  # 0 is mask value
     score = np.array([0.1, 0.2, 0.3])
 
     hist_iid = np.array([[1, 2, 3, 0], [1, 2, 3, 0], [1, 2, 0, 0]])
-    hist_igender = np.array([[1, 1, 2, 0], [2, 1, 1, 0], [2, 1, 0, 0]])
+    hist_cate_id = np.array([[1, 2, 2, 0], [1, 2, 2, 0], [1, 2, 0, 0]])
 
     behavior_length = np.array([3, 3, 2])
 
-    feature_dict = {'user': uid, 'gender': ugender, 'item_id': iid, 'cate_id': igender,
-                    'hist_item_id': hist_iid, 'hist_cate_id': hist_igender,
+    feature_dict = {'user': uid, 'gender': ugender, 'item_id': iid, 'cate_id': cate_id,
+                    'hist_item_id': hist_iid, 'hist_cate_id': hist_cate_id,
                     'pay_score': score, "seq_length": behavior_length}
 
     if use_neg:
         feature_dict['neg_hist_item_id'] = np.array([[1, 2, 3, 0], [1, 2, 3, 0], [1, 2, 0, 0]])
-        feature_dict['neg_hist_cate_id'] = np.array([[1, 1, 2, 0], [2, 1, 1, 0], [2, 1, 0, 0]])
+        feature_dict['neg_hist_cate_id'] = np.array([[1, 2, 2, 0], [1, 2, 2, 0], [1, 2, 0, 0]])
         feature_columns += [
             VarLenSparseFeat(SparseFeat('neg_hist_item_id', vocabulary_size=3 + 1, embedding_dim=8, embedding_name='item_id'),
                              maxlen=4, length_name="seq_length"),